int/cmd: renames the directory.
