+++ internal/cmd/helpers/internal/construct/bqparse.go
