+++ internal/cmd/helpers/internal/construct/bqvalidate.go
