+++ internal/cmd/helpers/internal/construct/bswrite.go
